+++ filter_manager/urls.py
@@ -6,4 +6,5 @@ urlpatterns = patterns('',
    url(r'^save_filter/$', views.save_filter,),
    url(r'^get_typeahead/$', views.get_typeahead,),
    url(r'^use_filter/$', views.use_filter,),
+   url(r'^get_filters_by_user/$', views.get_filters_by_user,),
 )